# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dme.Customer:
    properties:
      address1:
        type: string
      address2:
        type: string
      address3:
        type: string
      allowBackOrders:
        type: boolean
      allowTransactions:
        type: boolean
      altAddress1:
        type: string
      altAddress2:
        type: string
      altAddress3:
        type: string
      altCity:
        type: string
      altCountry:
        type: string
      altFirstName:
        type: string
      altLastName:
        type: string
      altPhone:
        type: string
      altState:
        type: string
      altZip:
        type: string
      balance:
        type: number
      boats:
        items:
          $ref: '#/definitions/dme.Boat'
        type: array
      categoryCodes:
        items:
          $ref: '#/definitions/dme.CategoryCode'
        type: array
      cellPhone:
        type: string
      city:
        type: string
      comments:
        type: string
      companyName:
        type: string
      country:
        type: string
      creditLimit:
        type: number
      customInformation:
        items:
          $ref: '#/definitions/dme.CustomInformation'
        type: array
      discount:
        type: number
      email:
        type: string
      emergencyContact:
        type: string
      emergencyPhone:
        type: string
      firstName:
        type: string
      id:
        type: string
      inactive:
        type: boolean
      inactiveDate:
        type: string
      integrationId:
        type: string
      invoices:
        items:
          $ref: '#/definitions/dme.InvoiceDetailed'
        type: array
      lastModified:
        type: string
      lastName:
        type: string
      name:
        type: string
      noCcSurcharge:
        type: boolean
      paymentTermsCode:
        type: string
      paymentTermsCodeDescription:
        type: string
      phone:
        type: string
      poRequired:
        type: boolean
      priceColumn:
        type: string
      prospectId:
        type: string
      sendMassEmail:
        type: boolean
      shipmentMethod:
        type: string
      shipmentMethodDescription:
        type: string
      state:
        type: string
      taxFlag:
        type: boolean
      taxId:
        type: string
      taxIdState:
        type: string
      taxSchema:
        type: string
      useAltAddress:
        type: boolean
      waitListEntries:
        items:
          $ref: '#/definitions/dme.WaitListEntry'
        type: array
      webId:
        type: string
      webPassword:
        type: string
      workPhone:
        type: string
      zip:
        type: string
    type: object
  dme.CustomerSearch:
    properties:
      address:
        type: string
      city:
        type: string
      companyName:
        type: string
      customerID:
        type: string
      email:
        type: string
      name:
        type: string
      prospectId:
        type: string
      state:
        type: string
      zip:
        type: string
    type: object
  dme.CustomerShort:
    properties:
      companyName:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  dme.CustomerUpdate:
    properties:
      address1:
        type: string
      address2:
        type: string
      address3:
        type: string
      altAddress1:
        type: string
      altAddress2:
        type: string
      altAddress3:
        type: string
      altCity:
        type: string
      altCountry:
        type: string
      altFirstName:
        type: string
      altLastName:
        type: string
      altPhone:
        type: string
      altState:
        type: string
      altZip:
        type: string
      cellPhone:
        type: string
      city:
        type: string
      companyName:
        type: string
      country:
        type: string
      emergencyContact:
        type: string
      emergencyPhone:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      name:
        type: string
      phone:
        type: string
      shipmentMethod:
        type: string
      shipmentMethodDescription:
        type: string
      state:
        type: string
      useAltAddress:
        type: boolean
      workPhone:
        type: string
      zip:
        type: string
    type: object
...
  requests.CreateUserRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      firstName:
        example: John
        type: string
      image:
        example: /images/profiles/johndoe.jpg
        type: string
      isActive:
        example: true
        type: boolean
      isSuperuser:
        example: false
        type: boolean
      lastName:
        example: Doe
        type: string
      marinaId:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      modules:
        $ref: '#/definitions/models.Modules'
      organizationId:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      password:
        example: SecureP@ssw0rd
        minLength: 12
        type: string
      permissions:
        $ref: '#/definitions/models.Permissions'
      phone:
        example: "+15551234567"
        type: string
      roleId:
        example: 550e8400-e29b-41d4-a716-446655440003
        type: string
      title:
        example: Manager
        type: string
      username:
        example: johndoe
        type: string
    required:
    - email
    - firstName
    - lastName
    - marinaId
    - organizationId
    - password
    - roleId
    type: object
  requests.CustomerCreateRequest:
    properties:
      address1:
        type: string
      address2:
        type: string
      address3:
        type: string
      altAddress1:
        type: string
      altAddress2:
        type: string
      altAddress3:
        type: string
      altCity:
        type: string
      altCountry:
        type: string
      altFirstName:
        type: string
      altLastName:
        type: string
      altPhone:
        type: string
      altState:
        type: string
      altZip:
        type: string
      cellPhone:
        type: string
      city:
        type: string
      companyName:
        type: string
      country:
        type: string
      emergencyContact:
        type: string
      emergencyPhone:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      name:
        type: string
      phone:
        type: string
      shipmentMethod:
        type: string
      shipmentMethodDescription:
        type: string
      state:
        type: string
      useAltAddress:
        type: boolean
      workPhone:
        type: string
      zip:
        type: string
    required:
    - lastName
    type: object
  requests.CustomerSettingsUpdateRequest:
    properties:
      customerId:
        type: string
      enablePortal:
        type: boolean
      marinaId:
        type: string
    required:
    - customerId
    - marinaId
    type: object
...
  responses.CustomerListResponse:
    properties:
      currentPage:
        example: 1
        type: integer
      data:
        items:
          $ref: '#/definitions/dme.Customer'
        type: array
      lastPage:
        example: 10
        type: integer
      perPage:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
  responses.CustomerListShortResponse:
    properties:
      currentPage:
        example: 1
        type: integer
      data:
        items:
          $ref: '#/definitions/dme.CustomerShort'
        type: array
      lastPage:
        example: 10
        type: integer
      perPage:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
  responses.CustomerResponse:
    properties:
      data:
        $ref: '#/definitions/dme.Customer'
    type: object
  responses.CustomerSearchResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dme.CustomerSearch'
        type: array
    type: object
  responses.CustomerSettingsResponse:
    properties:
      customerId:
        type: string
      enablePortal:
        type: boolean
      marinaId:
        type: string
    type: object
...
  responses.Error:
    description: Error response structure with optional validation details
    properties:
      details:
        items:
          $ref: '#/definitions/responses.ValidationError'
        type: array
      message:
        example: Validation failed
        type: string
    type: object
...
  responses.UserResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/responses.UserResponse'
      message:
        example: User profile retrieved successfully
        type: string
    type: object
  responses.ValidationError:
    description: Specific validation error for a single field
    properties:
      field:
        example: username
        type: string
      message:
        example: Username is required
        type: string
      tag:
        example: required
        type: string
      value:
        example: invalid_value
        type: string
    type: object
...
info:
  contact: {}
  description: This is a API Server.
  title: DockMaster Web API
  version: 0.0.1
paths:
...
  /customers/create:
    post:
      consumes:
      - application/json
      description: Creates a new customer
      parameters:
      - description: Customer information
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/requests.CustomerCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create customer
      tags:
      - Customers
  /customers/list:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of customers
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: List customers by page
      tags:
      - Customers
  /customers/list-short:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated short list of customers
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerListShortResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: List customers short
      tags:
      - Customers
  /customers/retrieve:
    get:
      consumes:
      - application/json
      description: Retrieves a customer by their ID
      parameters:
      - description: Customer ID
        in: query
        name: CustomerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Retrieve customer by ID
      tags:
      - Customers
  /customers/search:
    get:
      consumes:
      - application/json
      description: Searches for customers based on search string
      parameters:
      - description: Search string
        in: query
        name: SearchString
        required: true
        type: string
      - default: false
        description: Direct hit search
        in: query
        name: DirectHit
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Search customers
      tags:
      - Customers
  /customers/settings:
    get:
      consumes:
      - application/json
      description: Retrieves settings for a customer, creates with defaults if not
        found
      parameters:
      - description: Customer ID
        in: query
        name: CustomerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerSettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Get customer settings
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Updates settings for a customer
      parameters:
      - description: Customer settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/requests.CustomerSettingsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerSettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update customer settings
      tags:
      - Customers
  /customers/update:
    post:
      consumes:
      - application/json
      description: Updates a customer's information
      parameters:
      - description: Customer information
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/dme.CustomerUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update customer
      tags:
      - Customers
...
  /user:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created user
          schema:
            $ref: '#/definitions/responses.UserResponseWrapper'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/responses.Error'
      security:
      - ApiKeyAuth: []
      summary: Create user
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"